requirements.txt
Django==2.1.5
djangorestframework==3.9.1
pytz==2018.9
PyMySQL==0.9.3
Pillow==5.4.1
django-filter==2.1.0
